Cache the compiled xnb file after loading from raw asset files. It speeds up loading on Windows when used with FileSystemResolver.
There is one behavior change to the xna content processor. The nested build of an asset does not respect the processor parameters that are passed along with ContentProcessorContext.BuildAsset() method, instead it will be imported according to the side by side .importer settings file. This help avoids ambiguity where the same file might be processed with the same processor but with different parameters. The drawback to this design is that processor parameters that influence nested builds will no longer work, like the texture processor parameters specified in ModelProcessor. ContentProcessorContext.BuildAndLoadAsset is not affected by this behavior change.